+++ reference/separan/builtins.py
@@ -48,6 +48,7 @@ from .mail import MAIL_BUILTINS
 from .structured_data import STRUCTURED_DATA_BUILTINS
 from .embedded import EMBEDDED_BUILTINS
 from .network import NETWORK_BUILTINS
+from .network_services import NETWORK_SERVICE_BUILTINS
 
 
 MAX_TEXT_LENGTH = 1_048_576
@@ -798,7 +799,7 @@ BUILTINS = {
             , "cookie_error"
             , "db_connection_error", "db_auth_error", "db_query_error", "db_constraint_error", "db_timeout_error", "db_transaction_error", "db_driver_error"
             , "board_error", "pin_error", "peripheral_mapping_error", "embedded_backend_error"
-            , "network_error", "network_dns_error", "network_interface_error", "network_connection_error", "network_timeout_error", "network_limit_error", "network_closed_error", "network_protocol_error", "network_operation_unavailable", "network_address_error"
+            , "network_error", "network_dns_error", "network_interface_error", "network_connection_error", "network_timeout_error", "network_limit_error", "network_closed_error", "network_protocol_error", "network_operation_unavailable", "network_address_error", "network_service_error", "dhcp_server_error", "dns_server_error", "wifi_access_point_error"
         )),
-    ) + MATH_BUILTINS + (UtilityFunction("datetime", 1, 6, _datetime_construct, ("timezone",)),) + UTILITY_BUILTINS + PROCESS_BUILTINS + HTTP_BUILTINS + BYTES_BUILTINS + AUTH_BUILTINS + CRYPTO_BUILTINS + MAIL_BUILTINS + STRUCTURED_DATA_BUILTINS + COOKIE_BUILTINS + COOKIE_STORE_BUILTINS + SERVER_BUILTINS + DB_BUILTINS + EMBEDDED_BUILTINS + NETWORK_BUILTINS
+    ) + MATH_BUILTINS + (UtilityFunction("datetime", 1, 6, _datetime_construct, ("timezone",)),) + UTILITY_BUILTINS + PROCESS_BUILTINS + HTTP_BUILTINS + BYTES_BUILTINS + AUTH_BUILTINS + CRYPTO_BUILTINS + MAIL_BUILTINS + STRUCTURED_DATA_BUILTINS + COOKIE_BUILTINS + COOKIE_STORE_BUILTINS + SERVER_BUILTINS + DB_BUILTINS + EMBEDDED_BUILTINS + NETWORK_BUILTINS + NETWORK_SERVICE_BUILTINS
 }